Move TreeView controller to app/controllers
