+++ previews/primer/open_project/file_tree_view_preview/playground.html.erb
@@ -1,4 +1,4 @@
-<% data = TreeViewItemsController::TREE %>
+<% data = Primer::ViewComponents::TreeViewItemsController::TREE %>
 <%= render(Primer::OpenProject::FileTreeView.new) do |tree| %>
   <% populate.call(tree, data, { select_variant: select_variant, expanded: expanded }) %>
 <% end %>
