+++ previews/primer/open_project/tree_view_preview/loading_failure.html.erb
@@ -8,7 +8,7 @@
         <% icons.with_collapsed_icon(icon: :"file-directory-fill", color: :accent) %>
       <% end %>
 
-      <% sub_tree.with_loading_spinner(src: tree_view_items_path(loader: "spinner", fail: true)) %>
+      <% sub_tree.with_loading_spinner(src: primer_view_components.tree_view_items_path(loader: "spinner", fail: true)) %>
     <% end %>
   <% end %>
 </div>
